monitoring/requirements-reporting.txt
scikit-learn==1.2.0
imbalanced-learn==0.10.1 
dataclasses==0.6
pandas~=1.5.3
matplotlib==3.7.1
seaborn==0.12.2
numpy==1.22.4
pickleshare==0.7.5
pymongo==3.11.0
lightgbm==3.3.5
mlflow==2.3.0
evidently==0.1.51.dev0
gcsfs==2023.6.0
python-dotenv==1.0.0
google-cloud-storage==2.7.0